.old/back/requirements.txt
firebase-admin==6.2.0
Flask==2.3.2
gunicorn==20.1.0
plaid-python==13.0.0
python-dotenv==1.0.0